Remove unnecessary requires
The helper is already required through the `.rspec` file. Their presence
is unlikely to be a problem but can be annoying in some situations. For
example, if you introduce a syntax error in the helper, you will get it
printed hunders of times, one per file, since the require never succeeds
and it's retried everytime a new spec file is loaded.